requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10
langchain-huggingface==0.3.1
langchain-chroma==0.2.5
chromadb==0.4.22
streamlit==1.29.0
sentence-transformers==2.2.2
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
pypdf==3.17.4
python-multipart==0.0.6
fastapi==0.108.0
uvicorn==0.25.0